requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
all-packages==1.0.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
bidict==0.23.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cmake==3.31.0.1
colorama==0.4.6
distlib==0.3.8
distro==1.9.0
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.16.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-SocketIO==5.4.1
frozenlist==1.5.0
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.0
MarkupSafe==2.1.5
multidict==6.1.0
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==9.1.0
mysqlclient==2.2.6
numpy==2.1.3
openai==1.54.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
pillow==11.0.0
platformdirs==4.3.3
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
pypng==0.20220715.0
PyQRCode==1.2.1
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
qrcode==8.0
requests==2.32.3
setuptools==75.6.0
simple-websocket==1.1.0
sniffio==1.3.1
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.26.4
Werkzeug==3.1.3
wheel==0.45.0
wsproto==1.2.0
yarl==1.17.1
zipp==3.21.0
gunicorn